requirements.txt
# Agent Watch — Dependencies
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# AWS Bedrock
boto3>=1.34.0

# Neo4j
neo4j>=5.14.0

# Datadog (optional — falls back to local logging if not installed)
datadog>=0.47.0
ddtrace>=2.4.0
